+++ build.gradle
@@ -48,6 +48,9 @@ if (ENV.BUILD_ID) {
 group= "${config.group}" // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = "${config.name}"
 
+sourceCompatibility = 1.7
+targetCompatibility = 1.7
+
 minecraft {
     version = "${config.mc_version}-${config.forge_version}"
