+++ old/shapes/UtilSphericalFunctions.java
