+++ src/main/java/us/illyohs/azathoth/math/Vector3.java
@@ -26,7 +26,6 @@
 package us.illyohs.azathoth.math;
 
 import net.minecraft.util.EnumFacing;
-import us.illyohs.azathoth.world.WorldPos;
 
 
 public class Vector3 {
@@ -63,11 +62,11 @@ public class Vector3 {
     /**
      * Returns a difference vector such that reference + vector = destination
      */
-    public Vector3(WorldPos referencePoint, WorldPos worldPos){
-        x = worldPos.getX() - referencePoint.posX; 
-        y = worldPos.getY() - referencePoint.posY;
-        z = worldPos.getZ() - referencePoint.posZ;
-    }
+//    public Vector3(WorldPos referencePoint, WorldPos worldPos){
+//        x = worldPos.getX() - referencePoint.posX; 
+//        y = worldPos.getY() - referencePoint.posY;
+//        z = worldPos.getZ() - referencePoint.posZ;
+//    }
     
     public Vector3(EnumFacing face) {
         x = face.getFrontOffsetX();
